--- a/R/fine_tune.R
+++ b/R/fine_tune.R
@@ -28,7 +28,7 @@
 
     # We assume that classifySingleR() has already set up the backend.
     M <- .prep_for_parallel(exprs, BPPARAM)
-    S <- .prep_for_parallel(t(scores), BPPARAM)
+    S <- .cofragment_matrix(M, t(scores))
 
     bp.out <- bpmapply(Exprs=M, scores=S, FUN=fine_tune_label_de, 
         MoreArgs=list(References=references, quantile=quantile, tune_thresh=tune.thresh, marker_genes=de.info), 
@@ -52,11 +52,13 @@
 }
 
 #' @importFrom BiocParallel bpnworkers
-.prep_for_parallel <- function(mat, BPPARAM) {
+#' @importFrom DelayedArray colGrid getAutoBlockLength
+#' @importFrom BiocGenerics dims
+.prep_for_parallel <- function(mat, BPPARAM, use.grid=FALSE) {
     is.int <- !is.double(as.matrix(mat[0,0]))
     n_cores <- bpnworkers(BPPARAM)
 
-    if (n_cores==1L) {
+    if (n_cores==1L && !use.grid) {
         # Can't be bothered to template it twice at the C++ level,
         # as we'd have to have both int/numeric versions for the test and reference.
         if (is.int) {
@@ -67,9 +69,16 @@
 
     # Split the matrix *before* parallelization,
     # otherwise the full matrix gets serialized to all workers.
-    boundaries <- as.integer(seq(from = 1L, to = ncol(mat)+1L, length.out = n_cores + 1L)) 
-    out <- vector("list", n_cores)
+    if (!use.grid) {
+        boundaries <- as.integer(seq(from = 1L, to = ncol(mat)+1L, length.out = n_cores + 1L)) 
+    } else {
+        possible.block <- ceiling(ncol(mat) / n_cores) * nrow(mat)
+        allowed.block <- getAutoBlockLength(if (is.int) "integer" else "double")
+        grid <- colGrid(mat, block.length=min(possible.block, allowed.block))
+        boundaries <- c(1L, cumsum(dims(grid)[,2])+1L)
+    }
 
+    out <- vector("list", length(boundaries)-1L)
     for (i in seq_along(out)) {
         cur_start <- boundaries[i]
         cur_end <- boundaries[i+1]
@@ -81,4 +90,20 @@
     }
 
     out
+}
+
+.cofragment_matrix <- function(prepped, mat) 
+# Not breaking up `mat` in `.prep_for_parallel` as `use.grid=TRUE` means that
+# the choice of parallelization scheme is not purely driven by the number of
+# columns in `mat`. Also, `.prep_for_parallel` coerces to numeric, and
+# `mat` may not be.
+{
+    output <- vector("list", length(prepped)) 
+    counter <- 0L
+    for (i in seq_along(output)) {
+        N <- ncol(prepped[[i]])
+        output[[i]] <- mat[,counter + seq_len(N),drop=FALSE]
+        counter <- counter + N
+    }
+    output
 }