… NAMESPACE
… R/combineResults.R
… man/combineRecomputedResults.Rd
… man/trainSingleR.Rd
… src/RcppExports.cpp
… src/compute_scores.h
… src/scaled_ranks.h
… tests/testthat/test-combine.R
… tests/testthat/test-recomputed.R
… tests/testthat/test-train.R