A faster version of `combineRecomputedResults`, especially in parallelized scenarios. 

Technically speaking, it actually does more work, but in practice its overhead should be lower.