@dtm2451 2020-04-19
Sounds like @j-andrews7's results may not have been as positive, but for the same datasets as [last time](https://github.com/LTLA/SingleR/issues/94#issuecomment-605209834), I tend get a speed increase with nearly exactly the same results. (***n=1)

No. | # cells in test | Time (prev) | Time (*this*) | TimeDiff | % similarity
--- | --- | --- | --- | --- | ---
1 | 40,152 | 1131.123 sec | 1157.829 sec | +2.36% | 99.998 (1cell)
2 | 5,183 | 405.79 sec | 175.222 sec | -56.8% | 100
3 | 8,227 | 1461.29 sec | 1414.375 sec | -3.21% | 100
4 | 70,634 | 8382.182 sec | 7905.171 sec | -5.69% | 100