@j-andrews7 2020-04-19
# SingleR Recompute Testing

Comparing the original and new `recompute` methods.

For this data, I get a bit of a speedup like @dtm2451 saw when using the fine labels, but a slowdown when using the main labels. Reproduced twice, so I don't think it was just a fluke. As Dan saw, I get pretty similar results, and the changes seem to be between T cell subsets that are not super well differentiated and have a ton of labels in these references. 

**Summary:**

| Labels | Fine-tuning | Time (old) | Time (new) | Time (diff) | Similarity (%) | Similarity (n diff) |
|--------|-------------|------------|------------|-------------|----------------|---------------------|
| fine   | yes         | 424.44     | 361.02     | -14.9%      | 99.99          | 23                  |
| fine   | no          | 177.17     | 141.03     | -20.4%      | 99.99          | 13                  |
| main   | yes         | 152.27     | 243.52     | +59.9%      | 99.99          | 22                  |
| main   | no          | 83.42      | 164.53     | +97.2%      | 99.99          | 22                  |


## Full code

```{r libs, echo=FALSE}
suppressPackageStartupMessages(library(TENxPBMCData))
suppressPackageStartupMessages(library(scater))
suppressPackageStartupMessages(library(scran))
suppressPackageStartupMessages(library(SingleR))
```

```{r}
packageVersion("SingleR")
```
> 1.1.13

## Load and Pre-process Data

Basic filtering. Just a straight rip from the `scater` vignette.

```{r loadup, echo=FALSE}
# Load data
sce <- TENxPBMCData(dataset = "pbmc4k")
dim(sce)

# Basic filtering
per.cell <- perCellQCMetrics(sce, subsets=list(Mito=grep("mt-", rownames(sce))))
qc.stats <- quickPerCellQC(per.cell, percent_subsets="subsets_Mito_percent")
sce <- sce[,!qc.stats$discard]

keep <- nexprs(sce, byrow=TRUE) > 0
sce <- sce[keep, ]
sce <- logNormCounts(sce)
dim(sce)
```

> [1] 33694  4340
> [1] 19768  4311

## Dimensionality Reduction & Clustering

Again, just ripping off the `scran` vignette.

```{r dim, echo=FALSE}
# Get variable genes
dec <- modelGeneVar(sce)
top.hvgs <- getTopHVGs(dec, prop=0.1)

# Dim reductions
sce <- runPCA(sce, subset_row=top.hvgs)
sce <- runUMAP(sce)

output <- getClusteredPCs(reducedDim(sce))
npcs <- metadata(output)$chosen
reducedDim(sce, "PCAsub") <- reducedDim(sce, "PCA")[,1:npcs,drop=FALSE]
npcs

# Clustering
g <- buildSNNGraph(sce, use.dimred="PCAsub")
cluster <- igraph::cluster_walktrap(g)$membership
sce$cluster <- factor(cluster)

# Swap IDs to symbols
rownames(sce) <- rowData(sce)$Symbol_TENx
```

> 38

## Runs with old method

```{r refs, echo=false}
bp <- BlueprintEncodeData()
mona <- MonacoImmuneData()
dice <- DatabaseImmuneCellExpressionData()
dmap <- NovershternHematopoieticData()
hpca <- HumanPrimaryCellAtlasData()

refs <- list(BP=bp, Monaco=mona, DICE=dice, DMAP=dmap, HPCA=hpca)
fine.labels <- list(bp$label.fine, mona$label.fine, dice$label.fine, dmap$label.fine, 
    hpca$label.fine)
main.labels <- list(bp$label.main, mona$label.main, dice$label.main, 
    dmap$label.main, hpca$label.main)
```

```{r}
# Fine labels - single method
message("Recompute mode with fine labels and 'single' method:")
system.time(pred.fine.recomp <- SingleR(test = sce, ref = refs, labels = fine.labels, 
    method = "single"))
```

>    user  system elapsed 
 > 415.81    8.06  424.44 

```{r}
# Fine labels - single method - no fine tuning
message("Recompute mode with fine labels, 'single' method, and no fine-tuning:")
system.time(pred.fine.recomp.noft <- SingleR(test = sce, ref = refs, 
    labels = fine.labels, method = "single", fine.tune = FALSE))
```



```{r}
# Main labels - single method
message("Recompute mode with main labels and 'single' method:")
system.time(pred.main.recomp <- SingleR(test = sce, ref = refs, labels = main.labels, 
    method = "single"))
```

```{r}
# Main labels - single method - no fine tuning
message("Recompute mode with main labels, 'single' method, and no fine-tuning:")
system.time(pred.main.recomp.noft <- SingleR(test = sce, ref = refs, labels = main.labels, 
     method = "single", fine.tune = FALSE))
```

## Runs with new method

```{r}
# Terrible practice, don't do this.
detach(package:SingleR, unload = TRUE)
Sys.setenv(R_INSTALL_STAGED = FALSE)
devtools::install_github("LTLA/SingleR@faster-recompute", 
    lib = "C:/Users/Jared/Documents/R/dev")
suppressPackageStartupMessages(library(SingleR, 
    lib.loc = "C:/Users/Jared/Documents/R/dev/"))
packageVersion("SingleR")
```

```{r}
# Fine labels - single method
message("Recompute mode with fine labels and 'single' method:")
system.time(new.pred.fine.recomp <- SingleR(test = sce, ref = refs, 
    labels = fine.labels, method = "single"))
```

 > user  system elapsed 
 > 353.21    7.17  361.02 

```{r}
# Fine labels - single method - no fine tuning
message("Recompute mode with fine labels, 'single' method, and no fine-tuning:")
system.time(new.pred.fine.recomp.noft <- SingleR(test = sce, ref = refs, 
    labels = fine.labels, method = "single", fine.tune = FALSE))
```

> user  system elapsed 
> 133.72    7.05  141.03 

```{r}
# Main labels - single method
message("Recompute mode with main labels and 'single' method:")
system.time(new.pred.main.recomp <- SingleR(test = sce, ref = refs, 
    labels = main.labels, method = "single"))
```

> user  system elapsed 
> 238.90    3.94  243.52

```{r}
# Main labels - single method - no fine tuning
message("Recompute mode with main labels, 'single' method, and no fine-tuning:")
system.time(new.pred.main.recomp.noft <- SingleR(test = sce, ref = refs, 
    labels = main.labels, method = "single", fine.tune = FALSE))
```

> user  system elapsed 
> 160.68    3.59  164.53 

## Checking similarities

```{r}
identical(pred.fine.recomp$pruned.labels, new.pred.fine.recomp$pruned.labels)
message("Old:")
table(pred.fine.recomp$pruned.labels)
message("New:")
table(new.pred.fine.recomp$pruned.labels)
```

> Similarity:
> [1] 99.99466
> Number of dissimilar cells:
> [1] 23

```{r}
identical(pred.main.recomp$pruned.labels, new.pred.main.recomp$pruned.labels)
message("Old:")
table(pred.main.recomp$pruned.labels)
message("New:")
table(new.pred.main.recomp$pruned.labels)
```

> Similarity:
> [1] 99.99698
> Number of dissimilar cells:
> [1] 13

```{r}
identical(pred.fine.recomp.noft$pruned.labels, new.pred.fine.recomp.noft$pruned.labels)
message("Old:")
table(pred.fine.recomp.noft$pruned.labels)
message("New:")
table(new.pred.fine.recomp.noft$pruned.labels)
```

> Similarity:
> [1] 99.9949
> Number of dissimilar cells:
> [1] 22

```{r}
comp <- pred.main.recomp.noft$pruned.labels == new.pred.main.recomp.noft$pruned.labels
message("Similarity:")
100 - (length(comp[comp == FALSE]) / length(pred.main.recomp.noft$pruned.labels))
message("Number of dissimilar cells:")
length(comp[comp == FALSE])
```

> Similarity:
> [1] 99.9949
> Number of dissimilar cells:
> [1] 22